move sample from TTHAnalysis to RootTools
