+++ CMGTools/RootTools/python/samples/HLT_MC_SPRING15.txt
