+++ CMGTools/RootTools/python/samples/dumpTriggersFromFile.py
