+++ CMGTools/RootTools/python/samples/rareBg-caf.txt
