+++ CMGTools/RootTools/python/samples/triggers_13TeV_Spring15.py
