+++ CMGTools/RootTools/python/samples/triggers_8TeV.py
