+++ CMGTools/TTHAnalysis/cfg/run_susySingleSoftlepton_cfg.py
@@ -45,7 +45,7 @@ susyCoreSequence.insert(susyCoreSequence.index(ttHCoreEventAna),
                         ttHHeavyFlavourHadronAnalyzer)
 
 
-from CMGTools.TTHAnalysis.samples.samples_8TeV_v517 import triggers_1mu, triggers_1muHT, triggers_1eleHT # need to update the trigger MET pr HTMET?
+from CMGTools.RootTools.samples.samples_8TeV_v517 import triggers_1mu, triggers_1muHT, triggers_1eleHT # need to update the trigger MET pr HTMET?
 
 # Tree Producer
 treeProducer = cfg.Analyzer(
@@ -61,7 +61,7 @@ treeProducer = cfg.Analyzer(
 
 
 #-------- SAMPLES AND TRIGGERS -----------
-from CMGTools.TTHAnalysis.samples.samples_13TeV_CSA14 import *
+from CMGTools.RootTools.samples.samples_13TeV_CSA14 import *
 selectedComponents = [ SingleMu, DoubleElectron, TTHToWW_PUS14, DYJetsToLL_M50_PU20bx25, TTJets_PUS14 ]
