+++ CMGTools/TTHAnalysis/cfg/run_ttHLep8TeV_newNtpl_cfg.py
@@ -142,7 +142,7 @@ ttHEventAna = cfg.Analyzer(
     minJets25 = 0,
     )
 
-from CMGTools.TTHAnalysis.samples.samples_8TeV_v517 import triggers_mumu, triggers_ee, triggers_mue, triggers_1mu
+from CMGTools.RootTools.samples.samples_8TeV_v517 import triggers_mumu, triggers_ee, triggers_mue, triggers_1mu
 # Tree Producer
 treeProducer = cfg.Analyzer(
     #'ttHLepTreeProducerExample',
@@ -160,7 +160,7 @@ treeProducer = cfg.Analyzer(
 
 
 #-------- SAMPLES
-from CMGTools.TTHAnalysis.samples.samples_13TeV_CSA14 import * 
+from CMGTools.RootTools.samples.samples_13TeV_CSA14 import * 
 
 selectedComponents = [ SingleMu, DoubleElectron, TTHToWW_PUS14, DYJetsToLL_M50_PU20bx25, TTJets_PUS14 ]
